requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
imageio==2.22.4
    # via scikit-image
imutils==0.5.4
    # via -r requirements.in
networkx==2.8.8
    # via scikit-image
numpy==1.23.4
    # via
    #   -r requirements.in
    #   imageio
    #   opencv-contrib-python-headless
    #   pywavelets
    #   scikit-image
    #   scipy
    #   tifffile
opencv-contrib-python-headless==4.6.0.66
    # via -r requirements.in
packaging==21.3
    # via
    #   pytesseract
    #   scikit-image
pillow==9.3.0
    # via
    #   imageio
    #   pytesseract
    #   scikit-image
pyparsing==3.0.9
    # via packaging
pytesseract==0.3.10
    # via -r requirements.in
pywavelets==1.4.1
    # via scikit-image
scikit-image==0.19.3
    # via -r requirements.in
scipy==1.9.3
    # via scikit-image
tifffile==2022.10.10
    # via scikit-image
